better sliding window (#31)
Summary:
Pull Request resolved: https://github.com/facebookincubator/dns/pull/31

We only use aggregates from all the samples we store in sliding window.
When we are hit with 80k QPS and have sliding window of 60s, it means we store and process 4.8 million stat samples, each storing time and value itself. It's not cheap, both memory wise (golang's Time is relatively big, storing two 64bit ints for time itself plus a pointer to Location, so at least 24 bytes on 64bit system).

We then copy those slices of values many times, and eventually sort all those millions of samples to get min/max and avg.

**But what if I tell you that we don't need to store the samples at all?**

Instead, we store a ring buffer of N aggregates (which are very cheap thanks to Welford algorithm).
This allows us to keep 'sliding' aggregates while not bothering with removal of outdated samples.

Reviewed By: deathowl

Differential Revision: D45698916

fbshipit-source-id: 4973199d80beae2eb98957b9b5f568b4ac7e56cc