Implement basic NSID support
Summary:
This change adds a command-line flag to mirror the "whoami" TXT record output in the NSID response extension.  This allows debugging of responses using `dig +nsid ...`.  This is especially helpful when one nameserver IP is backed by a pool of instances that might be in different configurations, as it binds the debug info and the query answer together.

The existing "whoami" output is unchanged, except that (1) the ordering of the TXT records may have changed and (2) the cluster name is omitted when not applicable, instead of reporting as "notavailable".

The new NSID output conveys the same information as "whoami", but in "key=value" syntax.

The implementation has been restructured somewhat: the debug info construction has moved into a new `debuginfo` package, which is used by `whoami` and `nsid`.  This ensures that future extensions to the debug info are shared across both channels, and provides separation of content and presentation.

Reviewed By: deathowl

Differential Revision: D46865009

fbshipit-source-id: 498d2107f584b2d816a621377c922bf84b9fd4ee