significantly improve RocksDB compilation speed
Summary:
Our bottleneck was unsurprisingly in sorting the entire dataset, which of course is CPU bound and single-threaded.

But we don't need to do it like this.

The reason we did so was to be able to guarantee that same keys end up in same buckets, combined with RocksDB requirement for data being [pre-sorted when we use SST ingestion](https://github.com/facebook/rocksdb/wiki/Creating-and-Ingesting-SST-files).

Instead, with this change, we use fast hashing to put same keys into the same buckets, and sort each bucket separately, in parallel, so it's ready for ingestion.

NOTE: while RDB is happy with individually sorted SSTs, for real performance we need to either globally sort the data, or call compaction after the ingestion is done. This diff opted for sorting in-memory using merge sort, as this brings significant performance wins at small memory cost (~300mb on 50m keys).

The result is a significant speedup in DB compilation, by around 50s on our biggest internal shard with 54m records.

Reviewed By: deathowl

Differential Revision: D54191772

fbshipit-source-id: b328e0b4f84d4134cc415c6f1dee14357f24dfc3