fbdns: lift 2-byte restriction (reader part)
Summary:
The 2-byte format for location and map ID fields inherited from TinyDNS has several downsides:
the mnemonics are too short for even abbreviations to make sense
the binary encodings used by c10r and mercator are cryptic and not intelligible by humans which hurts debuggability
the related generating code has to hash longer strings to fit into the 2 bytes and thus having risk of collisions.

This diff is one of a few, bootstrapping the initiative: it revises the data structures and call paths in the consumer side of FBDNS.

Another diff, D55745224, handles the compilation part.

Reviewed By: leoleovich

Differential Revision: D55709600

fbshipit-source-id: bae4f986111960bfef63412a4262757ddda35e34