support rangepoints in self-test
Summary:
We have a self-test that makes sure we get responses from all subnets as defined by ipmaps.
It was a follow-up to the incident we had, and I wanna keep it working.
As we use rangepoints more now, I'm updating the selftest to support doing the same check with them.

Reviewed By: t3lurid3

Differential Revision: D57490993

fbshipit-source-id: 52d53ef9f5b7dc8fa787e9bfa3fa0b992ad780c8