Add e2e UDP benchmark
Summary:
This is intended to be a lightweight alternative to a full loadtest, to enable rapid assessment of performance-relevant changes.  It works by spraying a huge number of UDP query packets at the server to saturate the input queue, and counting how many are returned.  (This could be seen as estimating the throughput during an attack.)

On my devserver, this shows 32-36,000 RPS (highly variable between runs).

Reviewed By: abulimov

Differential Revision: D57574871

fbshipit-source-id: 91ab77bb4bf950064be796850acb26479970d723