Add a DoS Absorption Handler
Summary:
Under simulated DoS conditions, FBDNS's memory usage grows without bound, leading to OOM crashes.  We believe this occurs because the read thread creates a new Goroutine for every incoming query.  The more Goroutines are running, the longer each one takes to finish, leading to rapid growith in the number of simultaneous Goroutines.  Setting GOMEMLIMIT slows down the memory growth but does not stop it or prevent OOMs, because Goroutine creation is not blocked by the memory limit.

We would like FBDNS to degrade more gracefully under attack, by answering as many queries as it can and dropping the rest.

This diff adds an additional handler to the processing chain that limits the number of concurrently processed requests to 100 per CPU.  This should be enough to ensure that the CPU is fully utilized (even if some queries are blocked on I/O to RocksDB) without risking unbounded memory usage.

The design used here is similar to the suggestion in https://github.com/miekg/dns/pull/1068, but more principled and explicit.

Reviewed By: abulimov

Differential Revision: D57574870

fbshipit-source-id: 52b560fc30b48a1192831d5ab326aeaffa0918db