Back out "Add a DoS Absorption Handler"
Summary:
Original commit changeset: 52b560fc30b4

Original Phabricator Diff: D57574870

This diff appears to be broken in production (only for IPv4).  Our best guess is that some subtype of response is not correctly accounted for in the concurrency handler, leading to a leak in the number of counted pending queries.  Eventually the count reaches 100 and no more queries are accepted.

Reviewed By: mjdiffy

Differential Revision: D57987184

fbshipit-source-id: 8350d6ba0c5c9ffd2a5f621207d923576627509c