Unify throttle across IPs and add monitoring
Summary:
When throttling is enabled, this diff exposes a stat for the the percentage of the concurrency limit that we are using.  If this percentage reaches 100, the task is presumably throttled.

To produce meaningful statistics, this diff also unifies the throttle mechanism across all IP addresses.  A unified throttle is also more sensible for configuration.

Reviewed By: leoleovich

Differential Revision: D61037012

fbshipit-source-id: bd45863c1353b38e222007c20b2c88b8ee12e342