Fix throttle jamming bug
Summary:
This diff fixes a bug in the `throttle` module of fbdns.

The "throttle" module works by acquiring a semaphore when a query is received, and releasing it when the query is "retired".  If the semaphore is exhausted, new queries will be queued to avoid creating an excessive number of goroutines.  However, if queries can be discarded without releasing the semaphore, the throttle can become "jammed": the server always appears to be busy, so incoming queries are never dequeued.  This error was observed with invalid (i.e. unparseable) packets in the original implementation, requiring it to be reverted (D57987184).

However, there is another pathway for queries to be discarded early: MsgAcceptFunc.  This pathway discards queries that are syntactically valid but semantically inappropriate.

This diff adds interception of MsgAcceptFunc to release the semaphore.

Reviewed By: leoleovich

Differential Revision: D61052571

fbshipit-source-id: 183e63cf366c044cef14f5d0b35c208868013ff2