Stop session ticket refresh at server stop (#74)
Summary:
Pull Request resolved: https://github.com/facebook/dns/pull/74

FBDNS supports DNS over TLS, and periodically reloads its TLS Session Tickets from a shared cache to support TLS Session Resumption.  The goroutine responsible for this reload is supposed to run until a "quit" channel is closed.  However, this channel is never closed, so the goroutine, and the TLS config it maintains, will leak if the server is ever stopped.

In typical use, an FBDNS server only stops when the process is being shut down, so this is not a problem in production, but it is a correctness issue, and could become relevant in any environments where the server is being restarted in a long-lived process.

Reviewed By: abulimov

Differential Revision: D61350816