Deduplicate parseResult function
Summary:
This diff is 1/n to implement CNAME chasing.

In future diffs, I will be modifying the logic in the `parseResult` function. This function's code is duplicated across files `answer.go` and `answer_sorted.go`.

This diff extracts that into 1 function so the logic changes won't have to be duplicated across both files and also makes it easier to maintain in the future.

Reviewed By: bemasc

Differential Revision: D64012517

fbshipit-source-id: 9a5f1276a8296dd45f9c7c4deb81b517e0d647b6