Clean up variable use and logger abstraction (#84)
Summary:

As I was trying to abstract away logic into helper functions to enable CNAME chasing, I realized there was a redundancy in variables passed to the helper function. `state.Req` held the same value as `r`. This diff aims to clean that up by using the `state` variable where possible.

It also cleaned up the call to the `LogFailed` function which doesn't need to be passed in a `*dns.Msg`, `state.Req` already contains this value.

Differential Revision: D64629679