Pass in ecs to FindLocation
Summary:
This diff refactors FindLocation to enable combining of ECS scope prefix lengths while CNAME chasing.

### Problem
When we respond to an ECS query, we set the source scope to the mask of the Location ID. This tells the resolver that this query is valid for all clients in the subet.

Up until now, we only did 1 location look up so we set the source scope to the value of the location mask.

But while doing CNAME chasing, we do multiple location ID lookups by calling `FindLocation`. During the lookups, we might find that the location mask of one Location ID is more restrictive than the location mask of another. Due to that, we need to set the most restrictive mask for the ECS source scope.

To enable that, I refactored the ECS look up outside of `FindLocation`. The actual setting of this most restrictive mask will be done in the next diff.

Reviewed By: abulimov, bemasc

Differential Revision: D65793950

fbshipit-source-id: 569583a13c6eb8bfe78dfecaf1329b223952ebe5