Implement CNAME chasing
Summary:
This diff implements CNAME chasing to comply with [this section](https://datatracker.ietf.org/doc/html/rfc1034#section-3.6.2:~:text=CNAME%20RRs%20cause,just%20the%20CNAME.) of RFC 1034.

CNAME chasing is when we follow the chain of CNAMEs to resolve the record the client wants. For example, say the client requests an A record for `foo.example.com`. We don't have an A record for `foo.example.com` but we have a CNAME to `example.com`. So we restart the query for the A record at `example.com` and say we find an A record with value `1.1.1.1`. We return both of these records to the client.

The answer returned to the client would be:
```
foo.example.com		IN	CNAME 	    example.com
example.com		IN 	A		1.1.1.1
```
Notes:
- CNAME chasing occurs for all query types *except* CNAME and ANY
- CNAME chasing is location aware (i.e. it uses Location IDs to determine which CNAME RR to use)
- There are many error handling cases (one example is CNAME cycles should cause SERVFAIL)

Reviewed By: abulimov

Differential Revision: D64706464

fbshipit-source-id: 54f4414b6b35d18c50972dee123bdd9af7299d42