Add response time field
Summary:
Adding response time to requests logged using Scribe. It will allow us to do analysis and find interesting insights for our server such as what queries are taking the most time, how much time specific domains take, etc.

I'm also removing the calls to `LogFailed` in the `chaseCNAME` method because 1) it leads to duplicate logs since its being called once in there and then it gets called later in `ServeDNSWithRCODE`, 2) its purpose is for internal failure logging whereas `dns_logs` is used to log client requests, 3) We  log those failures to `glog` so if we need debugging info, it can be accessed through there. Logarithm can be used to increase the retention policy; I'm not sure if that's enabled or not, but its not a high priority right now.

Reviewed By: mjdiffy

Differential Revision: D71021104

fbshipit-source-id: cf8a2aa05f32599ecfe451ffde91d66062085b75