Update blazesym submodule (#99)
Summary:
Update the blazesym submodule to a version compatible with commit 63242850928c ("Update `blazesym` snapshot"). blazesym-c is now stable, so we may be able to go one step further and drop the submodule entirely and rely on distribution package (I don't have all the background why it exists), but I'll leave that for others to decide.

- [x] Added tests, if you've added code that should be tested
- [x] Updated the documentation, if you've changed APIs
- [x] Ensured the test suite passes
- [x] Made sure your code lints
- [x] Completed the Contributor License Agreement ("CLA")

Pull Request resolved: https://github.com/facebook/dns/pull/99

Reviewed By: abulimov

Differential Revision: D74242242

Pulled By: deathowl

fbshipit-source-id: a3a89347a1b12271224c5f27c242814bfe8f1581