build(deps): bump Go version to 1.25 (#109)
Summary:
<!--
Thank you for your interest in this project. Bugs filed and PRs submitted are appreciated!

Please make sure that you are familiar with and follow the Code of Conduct of this project which can be found at https://github.com/facebook/dns/blob/main/CODE_OF_CONDUCT.md

Also, please make sure you're familiar with and follow the instructions in the contributing guidelines which can  be found at https://github.com/facebook/dns/blob/main/CONTRIBUTING.md

-->

**What:**
Bump Go version to 1.25 for both the project and GH Actions.

**Why:**

The new version of Go introduces a handful of changes and optimizations that are worthwhile for adoption.

**How:**

**Risks:**

**Checklist**:

<!--
Have you done all of these things?
To check an item, place an "x" in the box like so: "- [x] Tests"
Add "N/A" to the end of each line that's irrelevant to your changes
-->

- [ ] Added tests, if you've added code that should be tested
- [ ] Updated the documentation, if you've changed APIs
- [x] Ensured the test suite passes
- [x] Made sure your code lints
- [x] Completed the Contributor License Agreement ("CLA")

Pull Request resolved: https://github.com/facebook/dns/pull/109

Reviewed By: deathowl

Differential Revision: D82114431

Pulled By: virtyaluk

fbshipit-source-id: b35a1afadb0113614cfe2ec32a97c3ff42b35ad7