update coredns to 1.12.4 (#111)
Summary:
<!--
Thank you for your interest in this project. Bugs filed and PRs submitted are appreciated!

Please make sure that you are familiar with and follow the Code of Conduct of this project which can be found at https://github.com/facebook/dns/blob/main/CODE_OF_CONDUCT.md

Also, please make sure you're familiar with and follow the instructions in the contributing guidelines which can  be found at https://github.com/facebook/dns/blob/main/CONTRIBUTING.md

-->

**What:**

**Why:**

**How:**

**Risks:**

**Checklist**:

<!--
Have you done all of these things?
To check an item, place an "x" in the box like so: "- [x] Tests"
Add "N/A" to the end of each line that's irrelevant to your changes
-->

- [ ] Added tests, if you've added code that should be tested
- [ ] Updated the documentation, if you've changed APIs
- [ ] Ensured the test suite passes
- [ ] Made sure your code lints
- [ ] Completed the Contributor License Agreement ("CLA")

Pull Request resolved: https://github.com/facebook/dns/pull/111

Reviewed By: pmazzini

Differential Revision: D82120043

Pulled By: deathowl

fbshipit-source-id: bf84fa680d67ab8c1b115256cacbd8cc297871d6